+++ src/main/java/com/logger/todo/CustomersRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CustomersRepository extends JpaRepository<Customers,Integer>{
 
-    void removeById(int id);
+    void removeById(long id);
 }
